Removal of unwanted code.
