+++ cp/cp_init.h
@@ -10,7 +10,6 @@
 
 #include <pcap.h>
 #include <byteswap.h>
-#include "cp_main.h"
 #include "cp_timer.h"
 
 #ifdef __cplusplus
@@ -18,18 +17,10 @@ extern "C" {
 #endif
 
 
-/**
- * @brief  : core identifiers for control plane threads
- */
-struct cp_params {
-	unsigned stats_core_id;
-};
-
 extern int s11_pcap_fd;
 extern int s5s8_sgwc_fd;
 extern int s5s8_pgwc_fd;
 extern int pfcp_sgwc_fd ;
-extern struct cp_params cp_params;
 
 /**
  * @brief  : Initializes Control Plane data structures, packet filters, and calls for the
@@ -47,13 +38,6 @@ init_cp(void);
  */
 uint8_t
 update_rstCnt(void);
-
-/**
- * @brief  : Add entry for recovery time into heartbeat recovery file
- * @param  : recov_time, recovery time
- * @return : Returns nothing
- */
-void recovery_time_into_file(uint32_t recov_time);
 
 /**
  * @brief  : starts the timer thread
