+++ cp/gtpv2/common/gtpv2_interface.h
@@ -9,7 +9,6 @@
 
 #include <stdint.h>
 #include "gtpv2_ie.h"
-#include "gtpv2_msg_struct.h"
 #include "ue.h"
 #include "sm_struct.h"
 #include "proc_struct.h"
@@ -99,35 +98,6 @@ process_bearer_resource_command(gtpv2c_header_t *gtpv2c_rx,
 
 
 /**
- * @brief  : parses gtpv2c message and populates parse_sgwc_s5s8_create_session_response_t structure
- * @param  : gtpv2c_rx
- *           buffer containing create bearer response message
- * @param  : csr
- *           data structure to contain required information elements from create
- *           create session response message
- * @return : - 0 if successful
- *           - > 0 if error occurs during packet filter parsing corresponds to 3gpp
- *           specified cause error value
- *           - < 0 for all other errors
- */
-
-int
-parse_sgwc_s5s8_create_session_response(gtpv2c_header_t *gtpv2c_rx,
-		sgwc_s5s8_create_session_response_t *csr);
-
-/**
- * @brief  : Handles processing of sgwc s5s8 create session response messages
- * @param  : gtpv2c_rx
- *           gtpc2c message reception  buffer containing the response message
- * @return : - 0 if successful
- *           - > 0 if error occurs during packet filter parsing corresponds to 3gpp
- *           specified cause error value
- *           - < 0 for all other errors
- */
-int
-process_sgwc_s5s8_create_session_response(gtpv2c_header_t *gtpv2c_rx);
-
-/**
  * @brief  : Handles processing of sgwc s11 create bearer response messages
  * @param  : gtpv2c_rx
  *           gtpc2c message reception  buffer containing the response message
@@ -171,31 +141,6 @@ process_pgwc_s5s8_create_session_request(gtpv2c_header_t *gtpv2c_rx,
 		struct in_addr *upf_ipv4, uint8_t proc);
 
 /**
- * @brief  : Handles the generation of sgwc s5s8 create session request messages
- * @param  : gtpv2c_s11_rx
- *           gtpc2c message reception  buffer containing s11 request message
- * @param  : gtpv2c_s5s8_tx
- *           gtpc2c message transmission buffer to contain s5s8 response message
- * @param  : sequence
- *           sequence number as described by clause 7.6 3gpp 29.274
- * @param  : pdn
- *           PDN Connection data structure pertaining to the session to be created
- * @param  : bearer
- *           Default EPS Bearer corresponding to the PDN Connection to be created
- * @param  : sgwu_fqdn
- *           SGWU fqdn to be sent to PGWC
- * @return : - 0 if successful
- *           - > 0 if error occurs during packet filter parsing corresponds to 3gpp
- *           specified cause error value
- *           - < 0 for all other errors
- */
-int
-gen_sgwc_s5s8_create_session_request(gtpv2c_header_t *gtpv2c_s11_rx,
-		gtpv2c_header_t *gtpv2c_s5s8_tx,
-		uint32_t sequence, pdn_connection_t *pdn,
-		eps_bearer_t *bearer, char *sgwu_fqdn);
-
-/**
  * @brief  : Handles the processing of delete bearer response messages received by the
  *           control plane.
  * @param  : gtpv2c_rx
@@ -207,28 +152,6 @@ gen_sgwc_s5s8_create_session_request(gtpv2c_header_t *gtpv2c_s11_rx,
  */
 int
 process_delete_bearer_response(gtpv2c_header_t *gtpv2c_rx);
-
-/**
- * @brief  : Handles the generation of sgwc s5s8 delete session request messages
- * @param  : gtpv2c_rx
- *           gtpv2c message buffer containing delete session request message
- * @param  : gtpv2c_tx
- *           gtpv2c message buffer to contain delete session response message
- * @param  : pgw_gtpc_del_teid
- *           Default pgw_gtpc_del_teid to be deleted on PGW
- * @param  : sequence
- *           sequence number as described by clause 7.6 3gpp 29.274
- * @param  : del_ebi
- *           Id of EPS Bearer to be deleted
- * @return : - 0 if successful
- *           - > 0 if error occurs during packet filter parsing corresponds to 3gpp
- *           specified cause error value
- *           - < 0 for all other errors
- */
-int
-gen_sgwc_s5s8_delete_session_request(gtpv2c_header_t *gtpv2c_rx,
-		gtpv2c_header_t *gtpv2c_tx, uint32_t pgw_gtpc_del_teid,
-		uint32_t sequence, uint8_t del_ebi);
 
 /**
  * @brief  : Handles the processing and reply of gtp echo requests received by the control plane
@@ -356,7 +279,6 @@ handle_unknown_msg(msg_info_t **msg, gtpv2c_header_t *gtpv2c_s11_rx);
 
 void process_s5s8_upd_bearer_response_pfcp_timeout(void *data);
 void process_pgwc_s5s8_delete_session_request_pfcp_timeout(void *data);
-void process_sgwc_s5s8_delete_session_request_pfcp_timeout(void *data);
 int handle_create_bearer_response_msg(msg_info_t **msg_p, gtpv2c_header_t *gtpv2c_rx);
 int handle_delete_bearer_response_msg(msg_info_t **msg_p, gtpv2c_header_t *gtpv2c_rx);
 
@@ -370,26 +292,10 @@ void* out_handler_gtp(void *data);
 uint32_t get_gtp_sequence(void);
 
 #ifdef FUTURE_NEEDS
-int handle_ddn_ack_msg(gtpv2c_header_t *gtpv2c_rx, msg_info_t *msg);
 int handle_update_bearer_response_msg(gtpv2c_header_t *gtpv2c_rx, msg_info_t *msg);
 int handle_delete_bearer_cmd_msg(gtpv2c_header_t *gtpv2c_rx, msg_info_t *msg);
-int handle_delete_pdn_conn_set_rsp(gtpv2c_header_t *gtpv2c_rx, msg_info_t *msg);
-int handle_update_pdn_conn_set_req(gtpv2c_header_t *gtpv2c_rx, msg_info_t *msg);
 int handle_update_pdn_conn_set_rsp(gtpv2c_header_t *gtpv2c_rx, msg_info_t *msg);
-int handle_pgw_restart_notf_ack(gtpv2c_header_t *gtpv2c_rx, msg_info_t *msg);
 #endif
-
-/**
- * @brief  : Fill Create Sess Request
- * @param  : cs_req, request structure to be filled
- * @param  : context, ue context info
- * @param  : ebi_index, index of bearer in bearer array
- * @return : Returns 0 in case of success , -1 otherwise
- */
-int
-fill_cs_request(create_sess_req_t *cs_req, ue_context_t *context,
-		uint8_t ebi_index);
-
 
 #ifdef __cplusplus
 }
