+++ cp/gtpv2/msg/delete_session_request.c
@@ -246,177 +246,8 @@ process_pgwc_s5s8_delete_session_request(del_sess_req_t *ds_req)
 	return 0;
 }
 
-void process_sgwc_s5s8_delete_session_request_pfcp_timeout(void *data)
-{
-    LOG_MSG(LOG_NEVER, "data = %p ", data);
-    return;
-}
-
-
-void
-fill_ds_request(del_sess_req_t *ds_req, ue_context_t *context,
-		 uint8_t ebi_index)
-{
-	int len = 0;
-	set_gtpv2c_header(&ds_req->header, 1,
-			GTP_DELETE_SESSION_REQ, context->pdns[ebi_index]->s5s8_pgw_gtpc_teid,
-			context->sequence);
-
-	set_ie_header(&ds_req->lbi.header, GTP_IE_EPS_BEARER_ID,
-			IE_INSTANCE_ZERO, sizeof(uint8_t));
-
-	set_ebi(&ds_req->lbi, IE_INSTANCE_ZERO,
-			context->eps_bearers[ebi_index]->eps_bearer_id);
-
-	if (context->uli.lai) {
-		ds_req->uli.lai = context->uli.lai;
-		ds_req->uli.lai2.lai_mcc_digit_2 = context->uli.lai2.lai_mcc_digit_2;
-		ds_req->uli.lai2.lai_mcc_digit_1 = context->uli.lai2.lai_mcc_digit_1;
-		ds_req->uli.lai2.lai_mnc_digit_3 = context->uli.lai2.lai_mnc_digit_3;
-		ds_req->uli.lai2.lai_mcc_digit_3 = context->uli.lai2.lai_mcc_digit_3;
-		ds_req->uli.lai2.lai_mnc_digit_2 = context->uli.lai2.lai_mnc_digit_2;
-		ds_req->uli.lai2.lai_mnc_digit_1 = context->uli.lai2.lai_mnc_digit_1;
-		ds_req->uli.lai2.lai_lac = context->uli.lai2.lai_lac;
-		len += sizeof(ds_req->uli.lai2);
-	}
-	if (context->uli.tai) {
-		ds_req->uli.tai = context->uli.tai;
-		ds_req->uli.tai2.tai_mcc_digit_2 = context->uli.tai2.tai_mcc_digit_2;
-		ds_req->uli.tai2.tai_mcc_digit_1 = context->uli.tai2.tai_mcc_digit_1;
-		ds_req->uli.tai2.tai_mnc_digit_3 = context->uli.tai2.tai_mnc_digit_3;
-		ds_req->uli.tai2.tai_mcc_digit_3 = context->uli.tai2.tai_mcc_digit_3;
-		ds_req->uli.tai2.tai_mnc_digit_2 = context->uli.tai2.tai_mnc_digit_2;
-		ds_req->uli.tai2.tai_mnc_digit_1 = context->uli.tai2.tai_mnc_digit_1;
-		ds_req->uli.tai2.tai_tac = context->uli.tai2.tai_tac;
-		len += sizeof(ds_req->uli.tai2);
-	}
-	if (context->uli.rai) {
-		ds_req->uli.rai = context->uli.rai;
-		ds_req->uli.rai2.ria_mcc_digit_2 = context->uli.rai2.ria_mcc_digit_2;
-		ds_req->uli.rai2.ria_mcc_digit_1 = context->uli.rai2.ria_mcc_digit_1;
-		ds_req->uli.rai2.ria_mnc_digit_3 = context->uli.rai2.ria_mnc_digit_3;
-		ds_req->uli.rai2.ria_mcc_digit_3 = context->uli.rai2.ria_mcc_digit_3;
-		ds_req->uli.rai2.ria_mnc_digit_2 = context->uli.rai2.ria_mnc_digit_2;
-		ds_req->uli.rai2.ria_mnc_digit_1 = context->uli.rai2.ria_mnc_digit_1;
-		ds_req->uli.rai2.ria_lac = context->uli.rai2.ria_lac;
-		ds_req->uli.rai2.ria_rac = context->uli.rai2.ria_rac;
-		len += sizeof(ds_req->uli.rai2);
-	}
-	if (context->uli.sai) {
-		ds_req->uli.sai = context->uli.sai;
-		ds_req->uli.sai2.sai_mcc_digit_2 = context->uli.sai2.sai_mcc_digit_2;
-		ds_req->uli.sai2.sai_mcc_digit_1 = context->uli.sai2.sai_mcc_digit_1;
-		ds_req->uli.sai2.sai_mnc_digit_3 = context->uli.sai2.sai_mnc_digit_3;
-		ds_req->uli.sai2.sai_mcc_digit_3 = context->uli.sai2.sai_mcc_digit_3;
-		ds_req->uli.sai2.sai_mnc_digit_2 = context->uli.sai2.sai_mnc_digit_2;
-		ds_req->uli.sai2.sai_mnc_digit_1 = context->uli.sai2.sai_mnc_digit_1;
-		ds_req->uli.sai2.sai_lac = context->uli.sai2.sai_lac;
-		ds_req->uli.sai2.sai_sac = context->uli.sai2.sai_sac;
-		len += sizeof(ds_req->uli.sai2);
-	}
-	if (context->uli.cgi) {
-		ds_req->uli.cgi = context->uli.cgi;
-		ds_req->uli.cgi2.cgi_mcc_digit_2 = context->uli.cgi2.cgi_mcc_digit_2;
-		ds_req->uli.cgi2.cgi_mcc_digit_1 = context->uli.cgi2.cgi_mcc_digit_1;
-		ds_req->uli.cgi2.cgi_mnc_digit_3 = context->uli.cgi2.cgi_mnc_digit_3;
-		ds_req->uli.cgi2.cgi_mcc_digit_3 = context->uli.cgi2.cgi_mcc_digit_3;
-		ds_req->uli.cgi2.cgi_mnc_digit_2 = context->uli.cgi2.cgi_mnc_digit_2;
-		ds_req->uli.cgi2.cgi_mnc_digit_1 = context->uli.cgi2.cgi_mnc_digit_1;
-		ds_req->uli.cgi2.cgi_lac = context->uli.cgi2.cgi_lac;
-	    ds_req->uli.cgi2.cgi_ci = context->uli.cgi2.cgi_ci;
-		len += sizeof(ds_req->uli.cgi2);
-	}
-	if (context->uli.ecgi) {
-		ds_req->uli.ecgi = context->uli.ecgi;
-		ds_req->uli.ecgi2.ecgi_mcc_digit_2 = context->uli.ecgi2.ecgi_mcc_digit_2;
-		ds_req->uli.ecgi2.ecgi_mcc_digit_1 = context->uli.ecgi2.ecgi_mcc_digit_1;
-		ds_req->uli.ecgi2.ecgi_mnc_digit_3 = context->uli.ecgi2.ecgi_mnc_digit_3;
-		ds_req->uli.ecgi2.ecgi_mcc_digit_3 = context->uli.ecgi2.ecgi_mcc_digit_3;
-		ds_req->uli.ecgi2.ecgi_mnc_digit_2 = context->uli.ecgi2.ecgi_mnc_digit_2;
-		ds_req->uli.ecgi2.ecgi_mnc_digit_1 = context->uli.ecgi2.ecgi_mnc_digit_1;
-		ds_req->uli.ecgi2.ecgi_spare = context->uli.ecgi2.ecgi_spare;
-	    ds_req->uli.ecgi2.eci = context->uli.ecgi2.eci;
-		len += sizeof(ds_req->uli.ecgi2);
-	}
-	if (context->uli.macro_enodeb_id) {
-		ds_req->uli.macro_enodeb_id = context->uli.macro_enodeb_id;
-		ds_req->uli.macro_enodeb_id2.menbid_mcc_digit_2 =
-			context->uli.macro_enodeb_id2.menbid_mcc_digit_2;
-		ds_req->uli.macro_enodeb_id2.menbid_mcc_digit_1 =
-			context->uli.macro_enodeb_id2.menbid_mcc_digit_1;
-		ds_req->uli.macro_enodeb_id2.menbid_mnc_digit_3 =
-			context->uli.macro_enodeb_id2.menbid_mnc_digit_3;
-		ds_req->uli.macro_enodeb_id2.menbid_mcc_digit_3 =
-			context->uli.macro_enodeb_id2.menbid_mcc_digit_3;
-		ds_req->uli.macro_enodeb_id2.menbid_mnc_digit_2 =
-			context->uli.macro_enodeb_id2.menbid_mnc_digit_2;
-		ds_req->uli.macro_enodeb_id2.menbid_mnc_digit_1 =
-			context->uli.macro_enodeb_id2.menbid_mnc_digit_1;
-		ds_req->uli.macro_enodeb_id2.menbid_spare =
-			context->uli.macro_enodeb_id2.menbid_spare;
-		ds_req->uli.macro_enodeb_id2.menbid_macro_enodeb_id =
-			context->uli.macro_enodeb_id2.menbid_macro_enodeb_id;
-		ds_req->uli.macro_enodeb_id2.menbid_macro_enb_id2 =
-			context->uli.macro_enodeb_id2.menbid_macro_enb_id2;
-		len += sizeof(ds_req->uli.macro_enodeb_id2);
-	}
-	if (context->uli.extnded_macro_enb_id) {
-		ds_req->uli.extnded_macro_enb_id = context->uli.extnded_macro_enb_id;
-		ds_req->uli.extended_macro_enodeb_id2.emenbid_mcc_digit_1 =
-			context->uli.extended_macro_enodeb_id2.emenbid_mcc_digit_1;
-		ds_req->uli.extended_macro_enodeb_id2.emenbid_mnc_digit_3 =
-			context->uli.extended_macro_enodeb_id2.emenbid_mnc_digit_3;
-		ds_req->uli.extended_macro_enodeb_id2.emenbid_mcc_digit_3 =
-			context->uli.extended_macro_enodeb_id2.emenbid_mcc_digit_3;
-		ds_req->uli.extended_macro_enodeb_id2.emenbid_mnc_digit_2 =
-			context->uli.extended_macro_enodeb_id2.emenbid_mnc_digit_2;
-		ds_req->uli.extended_macro_enodeb_id2.emenbid_mnc_digit_1 =
-			context->uli.extended_macro_enodeb_id2.emenbid_mnc_digit_1;
-		ds_req->uli.extended_macro_enodeb_id2.emenbid_smenb =
-			context->uli.extended_macro_enodeb_id2.emenbid_smenb;
-		ds_req->uli.extended_macro_enodeb_id2.emenbid_spare =
-			context->uli.extended_macro_enodeb_id2.emenbid_spare;
-		ds_req->uli.extended_macro_enodeb_id2.emenbid_extnded_macro_enb_id =
-			context->uli.extended_macro_enodeb_id2.emenbid_extnded_macro_enb_id;
-		ds_req->uli.extended_macro_enodeb_id2.emenbid_extnded_macro_enb_id2 =
-			context->uli.extended_macro_enodeb_id2.emenbid_extnded_macro_enb_id2;
-		len += sizeof(ds_req->uli.extended_macro_enodeb_id2);
-	}
-	len += 1;
-	set_ie_header(&ds_req->uli.header, GTP_IE_USER_LOC_INFO, IE_INSTANCE_ZERO,
-		len);
-}
-
 #endif
 #ifdef FUTURE_NEED
-int
-gen_sgwc_s5s8_delete_session_request(gtpv2c_header_t *gtpv2c_rx,
-		gtpv2c_header_t *gtpv2c_tx, uint32_t pgw_gtpc_del_teid,
-		uint32_t sequence, uint8_t del_ebi)
-{
-
-	gtpv2c_ie *current_rx_ie;
-	gtpv2c_ie *limit_rx_ie;
-
-	set_gtpv2c_teid_header(gtpv2c_tx, GTP_DELETE_SESSION_REQ,
-		    pgw_gtpc_del_teid, sequence);
-
-	FOR_EACH_GTPV2C_IE(gtpv2c_rx, current_rx_ie, limit_rx_ie)
-	{
-		if (current_rx_ie->type == GTP_IE_EPS_BEARER_ID &&
-				current_rx_ie->instance == IE_INSTANCE_ZERO) {
-			set_ebi_ie(gtpv2c_tx, IE_INSTANCE_ZERO, del_ebi);
-		} else if (current_rx_ie->type == GTP_IE_USER_LOC_INFO &&
-				current_rx_ie->instance == IE_INSTANCE_ZERO) {
-			set_ie_copy(gtpv2c_tx, current_rx_ie);
-		} else if (current_rx_ie->type == GTP_IE_INDICATION &&
-				current_rx_ie->instance == IE_INSTANCE_ZERO) {
-			set_ie_copy(gtpv2c_tx, current_rx_ie);
-		}
-	}
-
-	return 0;
-}
 
 /**
  * @brief  : Maintans gateway information
