+++ cp/pdn.c
@@ -133,41 +133,3 @@ get_new_bearer_id(pdn_connection_t *pdn_cntxt)
 	return pdn_cntxt->num_bearer;
 }
 
-uint8_t
-get_ue_state(uint32_t teid_key, uint8_t ebi_index)
-{
-	ue_context_t *context = NULL;
-	pdn_connection_t *pdn = NULL;
-	context  = (ue_context_t *)get_ue_context(teid_key);
-
-	if ( context == NULL) {
-		LOG_MSG(LOG_ERROR, "Entry not found for teid:%x...", teid_key);
-		return -1;
-	}
-	pdn = GET_PDN(context , ebi_index);
-	LOG_MSG(LOG_DEBUG, "Teid:%u, State:%s",
-			teid_key, get_state_string(pdn->state));
-	return pdn->state;
-}
-
-uint8_t
-update_ue_state(uint32_t teid_key, uint8_t state,  uint8_t ebi_index)
-{
-	ue_context_t *context = NULL;
-	pdn_connection_t *pdn = NULL;
-	context = (ue_context_t *)get_ue_context(teid_key);
-
-	if ( context == NULL) {
-		LOG_MSG(LOG_ERROR, "Failed to update UE State for Teid:%x...", 
-				teid_key);
-		return -1;
-	}
-	pdn = GET_PDN(context , ebi_index);
-	pdn->state = state;
-
-	LOG_MSG(LOG_DEBUG, "Change UE State for Teid:%u, State:%s",
-			teid_key, get_state_string(pdn->state));
-	return 0;
-
-}
-
