+++ cp/pfcp/common/pfcp_cp_association.c
@@ -13,7 +13,6 @@
 #include "pfcp_messages_decoder.h"
 #include "vepc_cp_dp_api.h"
 #include "cp_log.h"
-#include "cp_main.h"
 #include "pfcp.h"
 #include "spgw_config_struct.h"
 #include "gtpv2_error_rsp.h"
@@ -73,144 +72,6 @@ fill_pfcp_association_update_req(pfcp_assn_upd_req_t *pfcp_ass_update_req)
 
 }
 
-
-/* Fill pfd mgmt cstm ie */
-uint16_t
-set_pfd_contents(pfcp_pfd_contents_ie_t *pfd_conts, struct msgbuf1 *cstm_buf)
-{
-	pfd_conts->pfd_contents_spare = 0;
-	pfd_conts->pfd_contents_cp = 1;
-	/*pfd_conts->dn = 0;
-	pfd_conts->url = 0;
-	pfd_conts->fd = 0;
-	pfd_conts->pfd_contents_spare2 = 0x00;*/
-
-	if(pfd_conts->fd != 0){
-		pfd_conts->len_of_flow_desc = 0;
-		pfd_conts->flow_desc = 0;
-	}
-
-	if(pfd_conts->url != 0){
-		pfd_conts->length_of_url = 0;
-		pfd_conts->url2 = 0;
-	}
-
-	if(pfd_conts->dn != 0){
-		pfd_conts->len_of_domain_nm = 0;
-		pfd_conts->domain_name = 0;
-	}
-
-	if(pfd_conts->pfd_contents_cp != 0){
-		uint16_t struct_len = 0;
-		switch (cstm_buf->mtype) {
-			case MSG_SDF_CRE:
-			case MSG_ADC_TBL_CRE:
-			case MSG_PCC_TBL_CRE:
-			case MSG_SESS_TBL_CRE:
-			case MSG_MTR_CRE:
-				pfd_conts->cstm_pfd_cntnt = (uint8_t*)malloc(sizeof(struct cb_args_table));
-				/* Fill msg type */
-				struct_len = sprintf((char *)pfd_conts->cstm_pfd_cntnt, "%ld ",cstm_buf->mtype);
-				/* Fill cstm ie contents frome rule structure as string */
-				memcpy(pfd_conts->cstm_pfd_cntnt+struct_len, (uint8_t *)&cstm_buf->msg_union.msg_table,
-														sizeof(cstm_buf->msg_union.msg_table));
-				pfd_conts->len_of_cstm_pfd_cntnt = sizeof(cstm_buf->msg_union.msg_table) + struct_len;
-				break;
-
-			case MSG_EXP_CDR:
-				pfd_conts->cstm_pfd_cntnt = (uint8_t*)malloc(sizeof(struct msg_ue_cdr));
-				/* Fill msg type */
-				struct_len = sprintf((char *)pfd_conts->cstm_pfd_cntnt,
-												"%ld ",cstm_buf->mtype);
-				/* Fill cstm ie contents frome rule structure as string */
-				memcpy(pfd_conts->cstm_pfd_cntnt + struct_len,
-					  (uint8_t *)&cstm_buf->msg_union.ue_cdr, sizeof(struct msg_ue_cdr));
-				pfd_conts->len_of_cstm_pfd_cntnt = sizeof(struct msg_ue_cdr) + struct_len;
-				break;
-			case MSG_SDF_DES:
-			case MSG_ADC_TBL_DES:
-			case MSG_PCC_TBL_DES:
-			case MSG_SESS_TBL_DES:
-			case MSG_MTR_DES:
-				break;
-			case MSG_SDF_ADD:
-			case MSG_SDF_DEL:
-				pfd_conts->cstm_pfd_cntnt = (uint8_t*)malloc(sizeof(struct pkt_filter));
-				/* Fill msg type */
-				struct_len = sprintf((char *)pfd_conts->cstm_pfd_cntnt,
-													"%ld ",cstm_buf->mtype);
-				/* Fill cstm ie contents frome rule structure as string */
-				memcpy(pfd_conts->cstm_pfd_cntnt + struct_len,
-					  (uint8_t *)&cstm_buf->msg_union.pkt_filter_entry, sizeof(struct pkt_filter));
-				pfd_conts->len_of_cstm_pfd_cntnt = sizeof(struct pkt_filter)+struct_len;
-				break;
-			case MSG_ADC_TBL_ADD:
-			case MSG_ADC_TBL_DEL:
-				pfd_conts->cstm_pfd_cntnt = (uint8_t*)malloc(sizeof(struct adc_rules));
-				/* Fill msg type */
-				struct_len = sprintf((char *)pfd_conts->cstm_pfd_cntnt,
-												"%ld ",cstm_buf->mtype);
-				/* Fill cstm ie contents frome rule structure as string */
-				memcpy(pfd_conts->cstm_pfd_cntnt + struct_len,
-					  (uint8_t *)&cstm_buf->msg_union.adc_filter_entry, sizeof(struct adc_rules));
-				pfd_conts->len_of_cstm_pfd_cntnt = sizeof(struct adc_rules)+ struct_len;
-				break;
-			case MSG_PCC_TBL_ADD:
-			case MSG_PCC_TBL_DEL:
-				pfd_conts->cstm_pfd_cntnt = (uint8_t*)malloc(sizeof(struct pcc_rules));
-				/* Fill msg type */
-				struct_len = sprintf((char *)pfd_conts->cstm_pfd_cntnt,
-												"%ld ",cstm_buf->mtype);
-				/* Fill cstm ie contents frome rule structure as string */
-				memcpy(pfd_conts->cstm_pfd_cntnt + struct_len,
-					  (uint8_t *)&cstm_buf->msg_union.pcc_entry, sizeof(struct pcc_rules));
-				pfd_conts->len_of_cstm_pfd_cntnt = sizeof(struct pcc_rules) + struct_len;
-				break;
-			case MSG_SESS_CRE:
-			case MSG_SESS_MOD:
-			case MSG_SESS_DEL:
-				pfd_conts->cstm_pfd_cntnt = (uint8_t*)malloc(sizeof(struct session_info));
-				/* Fill msg type */
-				struct_len = sprintf((char *)pfd_conts->cstm_pfd_cntnt,
-												"%ld ",cstm_buf->mtype);
-				/* Fill cstm ie contents frome rule structure as string */
-				memcpy(pfd_conts->cstm_pfd_cntnt + struct_len,
-					  (uint8_t *)&cstm_buf->msg_union.sess_entry, sizeof(struct session_info));
-				pfd_conts->len_of_cstm_pfd_cntnt = sizeof(struct session_info) + struct_len;
-				break;
-			case MSG_MTR_ADD:
-			case MSG_MTR_DEL:
-				pfd_conts->cstm_pfd_cntnt = (uint8_t*)malloc(sizeof(struct mtr_entry));
-				/* Fill msg type */
-				struct_len = sprintf((char *)pfd_conts->cstm_pfd_cntnt,
-													"%ld ",cstm_buf->mtype);
-				/* Fill cstm ie contents frome rule structure as string */
-				memcpy(pfd_conts->cstm_pfd_cntnt + struct_len ,
-					  (uint8_t *)&cstm_buf->msg_union.mtr_entry, sizeof(struct mtr_entry));
-				pfd_conts->len_of_cstm_pfd_cntnt = sizeof(struct mtr_entry) + struct_len;
-				break;
-			case MSG_DDN_ACK:
-#if 0
-				pfd_conts->cstm_pfd_cntnt = malloc(sizeof(struct downlink_data_notification));
-				/* Fill msg type */
-				struct_len = sprintf((char *)pfd_conts->cstm_pfd_cntnt,
-																"%d ",cstm_buf->mtype);
-				/* Fill cstm ie contents frome rule structure as string */
-				memcpy(pfd_conts->cstm_pfd_cntnt + struct_len ,
-			          (uint8_t *)&cstm_buf->msg_union.mtr_entry, sizeof(struct downlink_data_notification));
-				pfd_conts->len_of_cstm_pfd_cntnt = sizeof(struct downlink_data_notification) + struct_len;
-#endif
-				break;
-			default:
-				LOG_MSG(LOG_ERROR, "build_dp_msg: Invalid msg type");
-				break;
-		}
-	}
-	/* set pfd contents header */
-	pfcp_set_ie_header(&pfd_conts->header, PFCP_IE_PFD_CONTENTS,
-			(pfd_conts->len_of_cstm_pfd_cntnt + 3));
-	return (pfd_conts->len_of_cstm_pfd_cntnt + 3);
-}
 
 /**
  * @brief  : This function fills in values to pfd context ie
