+++ cp/pfcp/common/pfcp_cp_util.c
@@ -6,44 +6,23 @@
 
 #include <sys/time.h>
 #include "pfcp_cp_util.h"
-#include "pfcp_cp_set_ie.h"
-#include "pfcp_messages.h"
-#include "cp_log.h"
-#include "spgw_config_struct.h"
-#include "cp_config_defs.h"
-#include "cp_io_poll.h"
-#include "util.h"
-#include "pfcp_cp_association.h"
-#include "spgw_cpp_wrapper.h"
-
-long
-uptime(void)
-{
-	struct sysinfo s_info;
-	int error = sysinfo(&s_info);
-	if(error != 0) {
-		LOG_MSG(LOG_DEBUG, "Error in uptime");
-	}
-	return s_info.uptime;
-}
 
 uint32_t
-current_ntp_timestamp(void) 
+current_ntp_timestamp(void)
 {
 
-	struct timeval tim;
-	uint8_t ntp_time[8] = {0};
-	uint32_t timestamp = 0;
+       struct timeval tim;
+       uint8_t ntp_time[8] = {0};
+       uint32_t timestamp = 0;
 
-	gettimeofday(&tim, NULL);
-	time_to_ntp(&tim, ntp_time);
+       gettimeofday(&tim, NULL);
+       time_to_ntp(&tim, ntp_time);
 
-	timestamp |= ntp_time[0] << 24 | ntp_time[1] << 16
-								| ntp_time[2] << 8 | ntp_time[3];
+       timestamp |= ntp_time[0] << 24 | ntp_time[1] << 16
+                                                               | ntp_time[2] << 8 | ntp_time[3];
 
-	return timestamp;
+       return timestamp;
 }
-
 void
 time_to_ntp(struct timeval *tv, uint8_t *ntp)
 {
