+++ cp/procedures/proc_initial_attach.c
@@ -904,9 +904,6 @@ fill_pdn_info(create_sess_req_t *csr, pdn_connection_t *pdn)
 
 	}
 
-	/* Note: s5s8_pgw_gtpc_teid updated by
-	 *                  * process_sgwc_s5s8_create_session_response (...)
-	 *                                   */
 	//pdn->s5s8_pgw_gtpc_teid = csr->s5s8pgw_pmip.teid_gre;
 
 	return 0;
@@ -1050,172 +1047,6 @@ fill_rule_and_qos_inform_in_pdn(pdn_connection_t *pdn)
     TAILQ_INSERT_TAIL(&pdn->policy.pending_pcc_rules, pcc_rule, next_pcc_rule);
 }
 
-
-
-
-/**
- * @brief  : parses gtpv2c message and populates parse_sgwc_s5s8_create_session_response_t structure
- * @param  : gtpv2c_rx
- *           buffer containing create bearer response message
- * @param  : csr
- *           data structure to contain required information elements from create
- *           create session response message
- * @return : - 0 if successful
- *           - > 0 if error occurs during packet filter parsing corresponds to 3gpp
- *             specified cause error value
- *           - < 0 for all other errors
- */
-
-int
-parse_sgwc_s5s8_create_session_response(gtpv2c_header_t *gtpv2c_rx,
-		struct parse_sgwc_s5s8_create_session_response_t *csr)
-{
-	gtpv2c_ie *current_ie;
-	gtpv2c_ie *current_group_ie;
-	gtpv2c_ie *limit_ie;
-	gtpv2c_ie *limit_group_ie;
-
-	FOR_EACH_GTPV2C_IE(gtpv2c_rx, current_ie, limit_ie)
-	{
-		if (current_ie->type == GTP_IE_BEARER_CONTEXT &&
-				current_ie->instance == IE_INSTANCE_ZERO) {
-			FOR_EACH_GROUPED_IE(current_ie, current_group_ie,
-					limit_group_ie)
-			{
-				if (current_group_ie->type == GTP_IE_EPS_BEARER_ID &&
-					current_group_ie->instance ==
-							IE_INSTANCE_ZERO) {
-					csr->bearer_context_to_be_created_ebi =
-					    IE_TYPE_PTR_FROM_GTPV2C_IE(uint8_t,
-							    current_group_ie);
-				} else if (current_group_ie->type ==
-						GTP_IE_BEARER_QLTY_OF_SVC &&
-						current_group_ie->instance ==
-							IE_INSTANCE_ZERO) {
-					csr->bearer_qos_ie = current_group_ie;
-				} else if (current_group_ie->type == GTP_IE_EPS_BEARER_LVL_TRAFFIC_FLOW_TMPL &&
-						current_group_ie->instance ==
-							IE_INSTANCE_ZERO) {
-					csr->bearer_tft_ie = current_group_ie;
-				} else if (current_group_ie->type == GTP_IE_FULLY_QUAL_TUNN_ENDPT_IDNT &&
-						current_group_ie->instance ==
-						IE_INSTANCE_ZERO) {
-					csr->s5s8_pgw_gtpu_fteid = current_group_ie;
-				}
-			}
-
-		} else if (current_ie->type == GTP_IE_FULLY_QUAL_TUNN_ENDPT_IDNT &&
-				current_ie->instance == IE_INSTANCE_ONE) {
-			csr->pgw_s5s8_gtpc_fteid =
-				IE_TYPE_PTR_FROM_GTPV2C_IE(fteid_ie, current_ie);
-		} else if (current_ie->type == GTP_IE_PDN_ADDR_ALLOC &&
-				current_ie->instance == IE_INSTANCE_ZERO) {
-			csr->pdn_addr_alloc_ie = current_ie;
-		} else if (current_ie->type == GTP_IE_APN_RESTRICTION &&
-				current_ie->instance == IE_INSTANCE_ZERO) {
-			csr->apn_restriction_ie = current_ie;
-		}
-	}
-
-	if (!csr->apn_restriction_ie
-		|| !csr->bearer_context_to_be_created_ebi
-		|| !csr->pgw_s5s8_gtpc_fteid) {
-		LOG_MSG(LOG_ERROR, "Dropping packet");
-		return GTPV2C_CAUSE_CONTEXT_NOT_FOUND;
-	}
-	return 0;
-}
-
-int
-gen_sgwc_s5s8_create_session_request(gtpv2c_header_t *gtpv2c_rx,
-		gtpv2c_header_t *gtpv2c_tx,
-		uint32_t sequence, pdn_connection_t *pdn,
-		eps_bearer_t *bearer, char *sgwu_fqdn)
-{
-
-	gtpv2c_ie *current_rx_ie;
-	gtpv2c_ie *limit_rx_ie;
-
-	set_gtpv2c_teid_header(gtpv2c_tx, GTP_CREATE_SESSION_REQ,
-		    0, sequence);
-
-	FOR_EACH_GTPV2C_IE(gtpv2c_rx, current_rx_ie, limit_rx_ie)
-	{
-		if (current_rx_ie->type == GTP_IE_BEARER_CONTEXT &&
-				current_rx_ie->instance == IE_INSTANCE_ZERO) {
-			gtpv2c_ie *bearer_context_group =
-				create_bearer_context_ie(gtpv2c_tx,
-			    IE_INSTANCE_ZERO);
-			add_grouped_ie_length(bearer_context_group,
-			    set_ebi_ie(gtpv2c_tx, IE_INSTANCE_ZERO,
-			    bearer->eps_bearer_id));
-			add_grouped_ie_length(bearer_context_group,
-				set_bearer_qos_ie(gtpv2c_tx, IE_INSTANCE_ZERO,
-				bearer));
-			add_grouped_ie_length(bearer_context_group,
-			    set_ipv4_fteid_ie(gtpv2c_tx, GTPV2C_IFTYPE_S5S8_SGW_GTPU,
-			    IE_INSTANCE_TWO, bearer->s5s8_sgw_gtpu_ipv4,
-			    htonl(bearer->s5s8_sgw_gtpu_teid)));
-		} else if (current_rx_ie->type == GTP_IE_FULLY_QUAL_TUNN_ENDPT_IDNT &&
-				current_rx_ie->instance == IE_INSTANCE_ONE) {
-			continue;
-		} else if (current_rx_ie->type == GTP_IE_FULLY_QUAL_TUNN_ENDPT_IDNT &&
-				current_rx_ie->instance == IE_INSTANCE_ZERO) {
-			set_ipv4_fteid_ie(gtpv2c_tx, GTPV2C_IFTYPE_S5S8_SGW_GTPC,
-				IE_INSTANCE_ZERO,
-				pdn->s5s8_sgw_gtpc_ipv4, htonl(pdn->s5s8_sgw_gtpc_teid));
-		} else if (current_rx_ie->type == GTP_IE_ACC_PT_NAME &&
-				current_rx_ie->instance == IE_INSTANCE_ZERO) {
-			set_ie_copy(gtpv2c_tx, current_rx_ie);
-		} else if (current_rx_ie->type == GTP_IE_APN_RESTRICTION &&
-				current_rx_ie->instance == IE_INSTANCE_ZERO) {
-			set_ie_copy(gtpv2c_tx, current_rx_ie);
-		} else if (current_rx_ie->type == GTP_IE_IMSI &&
-				current_rx_ie->instance == IE_INSTANCE_ZERO) {
-			set_ie_copy(gtpv2c_tx, current_rx_ie);
-		} else if (current_rx_ie->type == GTP_IE_AGG_MAX_BIT_RATE &&
-				current_rx_ie->instance == IE_INSTANCE_ZERO) {
-			set_ie_copy(gtpv2c_tx, current_rx_ie);
-		} else if (current_rx_ie->type == GTP_IE_PDN_TYPE &&
-				current_rx_ie->instance == IE_INSTANCE_ZERO) {
-			set_ie_copy(gtpv2c_tx, current_rx_ie);
-		} else if (current_rx_ie->type == GTP_IE_CHRGNG_CHAR &&
-				current_rx_ie->instance == IE_INSTANCE_ZERO) {
-			set_ie_copy(gtpv2c_tx, current_rx_ie);
-		} else if (current_rx_ie->type == GTP_IE_INDICATION &&
-				current_rx_ie->instance == IE_INSTANCE_ZERO) {
-			continue;
-		} else if (current_rx_ie->type == GTP_IE_MBL_EQUIP_IDNTY &&
-				current_rx_ie->instance == IE_INSTANCE_ZERO) {
-			continue;
-		} else if (current_rx_ie->type == GTP_IE_MSISDN &&
-				current_rx_ie->instance == IE_INSTANCE_ZERO) {
-			set_ie_copy(gtpv2c_tx, current_rx_ie);
-		} else if (current_rx_ie->type == GTP_IE_USER_LOC_INFO &&
-				current_rx_ie->instance == IE_INSTANCE_ZERO) {
-			set_ie_copy(gtpv2c_tx, current_rx_ie);
-		} else if (current_rx_ie->type == GTP_IE_SERVING_NETWORK &&
-				current_rx_ie->instance == IE_INSTANCE_ZERO) {
-			set_ie_copy(gtpv2c_tx, current_rx_ie);
-		} else if (current_rx_ie->type == GTP_IE_RAT_TYPE &&
-				current_rx_ie->instance == IE_INSTANCE_ZERO) {
-			set_ie_copy(gtpv2c_tx, current_rx_ie);
-		} else if (current_rx_ie->type == GTP_IE_SELECTION_MODE &&
-				current_rx_ie->instance == IE_INSTANCE_ZERO) {
-			set_ie_copy(gtpv2c_tx, current_rx_ie);
-		} else if (current_rx_ie->type == GTP_IE_PDN_ADDR_ALLOC &&
-				current_rx_ie->instance == IE_INSTANCE_ZERO) {
-			set_ie_copy(gtpv2c_tx, current_rx_ie);
-		} else if (current_rx_ie->type == GTP_IE_PROT_CFG_OPTS &&
-				current_rx_ie->instance == IE_INSTANCE_ZERO) {
-			set_ie_copy(gtpv2c_tx, current_rx_ie);
-		}
-	}
-
-	set_fqdn_ie(gtpv2c_tx, sgwu_fqdn);
-
-	return 0;
-}
 /**
  * @brief  : Table 7.2.1-1: Information Elements in a Create Session Request -
  *           incomplete list
@@ -1611,190 +1442,6 @@ set_pgwc_s5s8_create_session_response(gtpv2c_header_t *gtpv2c_tx,
 	}
 }
 
-//int
-//process_sgwc_s5s8_create_session_response(gtpv2c_header_t *gtpv2c_rx)
-//{
-//	int ret;
-//	ue_context *context = NULL;
-//	pdn_connection_t *pdn = NULL;
-//	eps_bearer_t *bearer = NULL;
-//	static uint32_t process_sgwc_s5s8_cs_rsp_cnt;
-//	static uint32_t process_spgwc_s11_cs_res_cnt;
-//
-//	struct resp_info *resp = NULL;
-//	pfcp_sess_mod_req_t pfcp_sess_mod_req = {0};
-//	//struct parse_sgwc_s5s8_create_session_response_t create_s5s8_session_response = {0};
-//	//ret = parse_sgwc_s5s8_create_session_response(gtpv2c_rx,
-//	//		&create_s5s8_session_response);
-//
-//	create_sess_rsp_t cs_rsp = {0};
-//	ret = decode_create_sess_rsp((uint8_t *)gtpv2c_rx, &cs_rsp);
-//	if (!ret)
-//		return ret;
-//
-//	uint8_t ebi_index =
-//		/**create_s5s8_session_response.bearer_context_to_be_created_ebi - 5;*/
-//		cs_rsp.bearer_contexts_created.eps_bearer_id.ebi_ebi - 5;
-//	//gtpv2c_rx->teid.has_teid.teid = ntohl(gtpv2c_rx->teid.has_teid.teid);
-//
-//	/* s11_sgw_gtpc_teid= s5s8_sgw_gtpc_teid =
-//	 * key->ue_context_by_fteid_hash */
-//	ret = r_t_e_hash_lookup_data(ue_context_by_fteid_hash,
-//	    (const void *) &gtpv2c_rx->teid.has_teid.teid,
-//	    (void **) &context);
-//
-//	LOG_MSG(LOG_DEBUG, "NGIC- create_s5s8_session.c::"
-//			"\n\tprocess_sgwc_s5s8_create_session_response"
-//			"\n\tprocess_sgwc_s5s8_cs_rsp_cnt= %u;"
-//			"\n\tgtpv2c_rx->teid.has_teid.teid= %X",
-//			process_sgwc_s5s8_cs_rsp_cnt,
-//			gtpv2c_rx->teid.has_teid.teid);
-//	if (ret < 0 || !context)
-//		return GTPV2C_CAUSE_CONTEXT_NOT_FOUND;
-//
-//	pdn = context->pdns[ebi_index];
-//	{
-//		pdn->apn_restriction =/* *IE_TYPE_PTR_FROM_GTPV2C_IE(uint8_t,
-//		    //create_s5s8_session_response.apn_restriction_ie);*/
-//			cs_rsp.apn_restriction.rstrct_type_val;
-//		struct in_addr ip;
-//		/*ip = get_ipv4_paa_ipv4(
-//					create_s5s8_session_response.pdn_addr_alloc_ie);*/
-//		ip =*(struct in_addr *) cs_rsp.paa.pdn_addr_and_pfx;
-//
-//		pdn->ipv4.s_addr = htonl(ip.s_addr);
-//
-//	//	pdn->s5s8_pgw_gtpc_ipv4.s_addr =
-//	//			htonl(create_s5s8_session_response.
-//	//			pgw_s5s8_gtpc_fteid->ip_u.ipv4.s_addr);
-//		pdn->s5s8_pgw_gtpc_ipv4.s_addr =
-//			cs_rsp.pgw_s5s8_s2as2b_fteid_pmip_based_intfc_or_gtp_based_ctl_plane_intfc.ipv4_address;
-//		/* Note: s5s8_pgw_gtpc_teid updated by
-//		 * create_s5s8_session_response.pgw_s5s8_gtpc_fteid...
-//		 */
-//	//	pdn->s5s8_pgw_gtpc_teid =
-//	//			htonl(create_s5s8_session_response.
-//	//			pgw_s5s8_gtpc_fteid->fteid_ie_hdr.teid_or_gre);
-//
-//		pdn->s5s8_pgw_gtpc_teid =
-//			htonl(cs_rsp.pgw_s5s8_s2as2b_fteid_pmip_based_intfc_or_gtp_based_ctl_plane_intfc.teid_gre_key);
-//
-//	}
-//	bearer = context->eps_bearers[ebi_index];
-//	{
-//		/* TODO: Implement TFTs on default bearers
-//		if (create_s5s8_session_response.bearer_tft_ie) {
-//		}
-//		*/
-//		/* TODO: Implement PGWC S5S8 bearer QoS */
-//		if (cs_rsp.bearer_contexts_created.bearer_lvl_qos.header.len) {
-//			//bearer->qos = *IE_TYPE_PTR_FROM_GTPV2C_IE(bearer_qos_ie,
-//				//create_s5s8_session_response.bearer_qos_ie);
-//				bearer->qos.qci = cs_rsp.bearer_contexts_created.bearer_lvl_qos.qci;
-//				bearer->qos.ul_mbr = cs_rsp.bearer_contexts_created.bearer_lvl_qos.max_bit_rate_uplnk;
-//				bearer->qos.dl_mbr = cs_rsp.bearer_contexts_created.bearer_lvl_qos.max_bit_rate_dnlnk;
-//				bearer->qos.ul_gbr = cs_rsp.bearer_contexts_created.bearer_lvl_qos.guarntd_bit_rate_uplnk;
-//				bearer->qos.dl_gbr = cs_rsp.bearer_contexts_created.bearer_lvl_qos.guarntd_bit_rate_dnlnk;
-//				bearer->qos.arp.preemption_vulnerability = cs_rsp.bearer_contexts_created.bearer_lvl_qos.pvi;
-//				bearer->qos.arp.spare1 = cs_rsp.bearer_contexts_created.bearer_lvl_qos.spare2;
-//				bearer->qos.arp.priority_level = cs_rsp.bearer_contexts_created.bearer_lvl_qos.pl;
-//				bearer->qos.arp.preemption_capability =cs_rsp.bearer_contexts_created.bearer_lvl_qos.pci;
-//				bearer->qos.arp.spare2 = cs_rsp.bearer_contexts_created.bearer_lvl_qos.spare3;
-//		}
-//	//	bearer->s5s8_pgw_gtpu_ipv4.s_addr =
-//	//			htonl(IE_TYPE_PTR_FROM_GTPV2C_IE(fteid_ie,
-//	//					create_s5s8_session_response.s5s8_pgw_gtpu_fteid)->
-//	//						ip_u.ipv4.s_addr);
-//	//	bearer->s5s8_pgw_gtpu_teid =
-//	//			IE_TYPE_PTR_FROM_GTPV2C_IE(fteid_ie,
-//	//					create_s5s8_session_response.s5s8_pgw_gtpu_fteid)->
-//	//						fteid_ie_hdr.teid_or_gre;
-//
-//	//	bearer->s5s8_pgw_gtpu_ipv4.s_addr = cs_rsp.bearer_contexts_created.s5s8_u_pgw_fteid.ipv4_address;
-//		bearer->s5s8_pgw_gtpu_ipv4.s_addr = cs_rsp.bearer_contexts_created.s1u_sgw_fteid.ipv4_address;
-//	//	bearer->s5s8_pgw_gtpu_teid = cs_rsp.bearer_contexts_created.s5s8_u_pgw_fteid.teid_gre_key;
-//		bearer->s5s8_pgw_gtpu_teid = cs_rsp.bearer_contexts_created.s1u_sgw_fteid.teid_gre_key;
-//		bearer->pdn = pdn;
-//	}
-//
-//	s11_mme_sockaddr.sin_addr.s_addr =
-//					htonl(context->s11_mme_gtpc_ipv4.s_addr);
-//
-//	LOG_MSG(LOG_DEBUG, "NGIC- create_s5s8_session.c::"
-//			"\n\tprocess_sgwc_s5s8_cs_rsp_cnt= %u;"
-//			"\n\tprocess_spgwc_s11_cs_res_cnt= %u;"
-//			"\n\tue_ip= pdn->ipv4= %s;"
-//			"\n\tpdn->s5s8_sgw_gtpc_ipv4= %s;"
-//			"\n\tpdn->s5s8_sgw_gtpc_teid= %X;"
-//			"\n\tbearer->s5s8_sgw_gtpu_ipv4= %s;"
-//			"\n\tbearer->s5s8_sgw_gtpu_teid= %X;"
-//			"\n\tpdn->s5s8_pgw_gtpc_ipv4= %s;"
-//			"\n\tpdn->s5s8_pgw_gtpc_teid= %X;"
-//			"\n\tbearer->s5s8_pgw_gtpu_ipv4= %s;"
-//			"\n\tbearer->s5s8_pgw_gtpu_teid= %X",
-//			process_sgwc_s5s8_cs_rsp_cnt++,
-//			process_spgwc_s11_cs_res_cnt++,
-//			inet_ntoa(pdn->ipv4),
-//			inet_ntoa(pdn->s5s8_sgw_gtpc_ipv4),
-//			pdn->s5s8_sgw_gtpc_teid,
-//			inet_ntoa(bearer->s5s8_sgw_gtpu_ipv4),
-//			bearer->s5s8_sgw_gtpu_teid,
-//			inet_ntoa(pdn->s5s8_pgw_gtpc_ipv4),
-//			pdn->s5s8_pgw_gtpc_teid,
-//			inet_ntoa(bearer->s5s8_pgw_gtpu_ipv4),
-//			bearer->s5s8_pgw_gtpu_teid);
-//
-//	uint32_t  seq_no = 0;
-//	seq_no = bswap_32(gtpv2c_rx->teid.has_teid.seq) ;
-//	seq_no = seq_no >> 8;
-//
-//	fill_pfcp_sess_mod_req(&pfcp_sess_mod_req, NULL, context, bearer, pdn);
-//	if(pfcp_sess_mod_req.create_pdr_count){
-//		pfcp_sess_mod_req.create_pdr[0].pdi.local_fteid.teid = htonl(bearer->s5s8_pgw_gtpu_teid) ;
-//		pfcp_sess_mod_req.create_pdr[0].pdi.local_fteid.ipv4_address = htonl(bearer->s5s8_pgw_gtpu_ipv4.s_addr) ;
-//		pfcp_sess_mod_req.create_pdr[0].pdi.ue_ip_address.ipv4_address = (pdn->ipv4.s_addr);
-//		pfcp_sess_mod_req.create_pdr[0].pdi.src_intfc.interface_value = SOURCE_INTERFACE_VALUE_ACCESS;
-//	}else if(pfcp_sess_mod_req.update_far_count){
-//		pfcp_sess_mod_req.update_far[0].upd_frwdng_parms.outer_hdr_creation.teid = bearer->s5s8_pgw_gtpu_teid;
-//		pfcp_sess_mod_req.update_far[0].upd_frwdng_parms.outer_hdr_creation.ipv4_address  = htonl(bearer->s5s8_pgw_gtpu_ipv4.s_addr) ;
-//		pfcp_sess_mod_req.update_far[0].upd_frwdng_parms.dst_intfc.interface_value = SOURCE_INTERFACE_VALUE_CORE;
-//	}
-//	pfcp_sess_mod_req.header.seid_seqno.has_seid.seq_no = gtpv2c_rx->teid.has_teid.seq ;
-//
-//	uint8_t pfcp_msg[512]={0};
-//	int encoded = encode_pfcp_sess_mod_req_t(&pfcp_sess_mod_req, pfcp_msg);
-//	pfcp_header_t *header = (pfcp_header_t *) pfcp_msg;
-//	header->message_len = htons(encoded - 4);
-//
-//
-//	if (pfcp_send(my_sock.sock_fd_pfcp, pfcp_msg, encoded, &context->upf_context->upf_sockaddr) < 0)
-//		LOG_MSG(LOG_ERROR, "Error in sending MBR to SGW-U. err_no: %i", errno);
-//	else
-//	{
-//	}
-//	/* Update UE State */
-//	context->state = PFCP_SESS_MOD_REQ_SNT_STATE;
-//
-//	/* Lookup Stored the session information. */
-//	resp = get_sess_entry_seid(context->pdns[ebi_index]->seid);
-//	if (resp == NULL) {
-//		LOG_MSG(LOG_ERROR, "Failed to add response in entry in SM_HASH");
-//		return -1;
-//	}
-//
-//	/* Set create session response */
-//	resp->sequence = seq_no;
-//	//resp->eps_bearer_id =
-//	//		*create_s5s8_session_response.bearer_context_to_be_created_ebi;
-//	resp->eps_bearer_id = cs_rsp.bearer_contexts_created.eps_bearer_id.ebi_ebi;
-//	resp->s11_sgw_gtpc_teid = context->s11_sgw_gtpc_teid;
-//	resp->context = context;
-//	resp->msg_type = GTP_CREATE_SESSION_RSP;
-//	resp->state = PFCP_SESS_MOD_REQ_SNT_STATE;
-//
-//	return 0;
-//}
-
 int
 gen_ccr_request(proc_context_t *proc_context, uint8_t ebi_index, create_sess_req_t *csr)
 {
@@ -2059,4 +1706,3 @@ process_pfcp_sess_est_request_timeout(void *data)
     proc_context->handler(proc_context, msg); 
     return;
 }
-
