+++ cp/sm_structs_api.h
@@ -14,24 +14,6 @@
 #ifdef __cplusplus
 extern "C" {
 #endif
-/**
- * @brief  : Update UE state in UE Context.
- * @param  : teid_key, key to search context
- * @param  : state, new state to be updated
- * @param  : ebi_index, index of bearer id stored in array
- * @return : Returns 0 in case of success , -1 otherwise
- */
-uint8_t
-update_ue_state(uint32_t teid_key, uint8_t state, uint8_t ebi_index);
-
-/**
- * @brief  : Retrive UE state from UE Context.
- * @param  : teid_key, key for search
- * @param  : ebi_index, index of bearer id stored in array
- * @return : Returns 0 in case of success , -1 otherwise
- */
-uint8_t
-get_ue_state(uint32_t teid_key ,uint8_t ebi_index);
 
 /**
  * Retrive ue context entry from Bearer table,using sgwc s5s8 teid.
